+++ src/SFA.DAS.EmployerFinance.MessageHandlers/EventHandlers/CreatedPaymentEventHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using NServiceBus;
-using SFA.DAS.EmployerAccounts.Events.Messages;
 using SFA.DAS.EmployerFinance.Messages.Events;
 using SFA.DAS.Messaging.Interfaces;
